Remove model rules from publishing plugin
The publishing plugin's model rules are only
necessary for backwards compatiblity with builds
that define publications using model rules.

The rules are now initialized lazily so that the
publishing plugin does not trigger model bridging.
